move `SmtLeaf` to submodule
