+++ src/merkle/smt/full/tests.rs
@@ -1,3 +1,5 @@
+use winter_utils::{Deserializable, Serializable};
+
 use super::*;
 use crate::{
     merkle::{EmptySubtreeRoots, MerkleStore},
